feat: Initial edit for next-edit completions
I'm not very happy with the prompt (I think it needs to be workshopped), and I
want to rework a bunch of the actual `next-edit.ts`, but it somewhat works.

One of the cool things here is that `scripts/preview-next-edit.ts` is a script
that evaluates the quality of a given set of next edits, so it could be used for
benchmarking different models for both time and quality.